+++ backend/integration_tests/src/register_user_tests.rs
@@ -20,7 +20,7 @@ fn register_users() {
         users[0].principal,
         canister_ids.user_index,
         &user_index_canister::users::Args {
-            user_groups: vec![user_index_canister::users::UserGroup {
+            user_groups: vec![user_index_canister::users_v2::UserGroup {
                 users: users.iter().map(|u| u.user_id).collect(),
                 updated_since: 0,
             }],
@@ -52,6 +52,7 @@ fn register_user_with_duplicate_username_appends_suffix() {
             canister_ids.local_user_index,
             &local_user_index_canister::register_user::Args {
                 username: username.clone(),
+                display_name: None,
                 referral_code: None,
                 public_key,
             },
@@ -70,7 +71,7 @@ fn register_user_with_duplicate_username_appends_suffix() {
         first_user_principal.unwrap(),
         canister_ids.user_index,
         &user_index_canister::users::Args {
-            user_groups: vec![user_index_canister::users::UserGroup {
+            user_groups: vec![user_index_canister::users_v2::UserGroup {
                 users: user_ids,
                 updated_since: 0,
             }],