+++ frontend/openchat-sw/rollup.config.mjs
@@ -70,10 +70,7 @@ export default {
     },
     plugins: [
         production && clean(),
-        typescript({
-            sourceMap: !production,
-            inlineSources: !production,
-        }),
+        typescript(),
 
         // In dev mode, call `npm run start` once
         // the bundle has been generated